+++ src/main/java/woowacourse/shoppingcart/application/CustomerService.java
@@ -1,0 +1,62 @@
+package woowacourse.shoppingcart.application;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import woowacourse.shoppingcart.dao.CustomerDao;
+import woowacourse.shoppingcart.domain.Customer;
+import woowacourse.shoppingcart.dto.CustomerRequest;
+import woowacourse.shoppingcart.dto.CustomerResponse;
+import woowacourse.shoppingcart.exception.InvalidCustomerException;
+
+@Service
+@Transactional
+public class CustomerService {
+
+    private final CustomerDao customerDao;
+
+    public CustomerService(CustomerDao customerDao) {
+        this.customerDao = customerDao;
+    }
+
+    @Transactional(readOnly = true)
+    public boolean isExistEmail(String email) {
+        return customerDao.existByEmail(email);
+    }
+
+    public CustomerResponse save(CustomerRequest customerRequest) {
+        validateDuplicationEmail(customerRequest);
+        Customer customer = customerRequest.createCustomer();
+        customerDao.save(customer);
+        return CustomerResponse.from(customer);
+    }
+
+    private void validateDuplicationEmail(CustomerRequest customerRequest) {
+        if (isExistEmail(customerRequest.getEmail())) {
+            throw new IllegalArgumentException("중복된 email 입니다.");
+        }
+    }
+
+    @Transactional(readOnly = true)
+    public CustomerResponse find(Long customerId) {
+        checkExistById(customerId);
+        Customer customer = customerDao.findById(customerId);
+        return CustomerResponse.from(customer);
+    }
+
+    private void checkExistById(Long customerId) {
+        if (!customerDao.existById(customerId)) {
+            throw new InvalidCustomerException();
+        }
+    }
+
+    public void update(Long customerId, CustomerRequest customerRequest) {
+        checkExistById(customerId);
+        Customer customer = customerRequest.createCustomer(customerId);
+        customerDao.update(customer);
+    }
+
+    public void delete(Long customerId) {
+        checkExistById(customerId);
+        customerDao.deleteById(customerId);
+    }
+}
